refactor: SubCategoryMenu 타입 분리
